Split apiFactories by domain
