+++ src/background/starterBlueprints.test.ts
@@ -33,14 +33,14 @@ import {
 import { uuidv4 } from "@/types/helpers";
 import { type RegistryId } from "@/types/registryTypes";
 import { type OutputKey } from "@/types/runtimeTypes";
-import { sanitizedAuthFactory } from "@/testUtils/factories/apiFactories";
 import { extensionFactory } from "@/testUtils/factories/extensionFactories";
 import {
   extensionPointConfigFactory,
   getRecipeWithBuiltInServiceAuths,
   recipeFactory,
 } from "@/testUtils/factories/recipeFactories";
-import { organizationFactory } from "@/testUtils/factories/authFactories";
+import { userOrganizationFactory } from "@/testUtils/factories/authFactories";
+import { sanitizedAuthFactory } from "@/testUtils/factories/serviceFactories";
 
 const axiosMock = new MockAdapter(axios);
 
@@ -95,7 +95,7 @@ describe("installStarterBlueprints", () => {
       .onGet("/api/services/shared/?meta=1")
       .reply(200, [
         sanitizedAuthFactory(),
-        sanitizedAuthFactory({ organization: organizationFactory() }),
+        sanitizedAuthFactory({ organization: userOrganizationFactory() }),
         sanitizedAuthFactory({ user: uuidv4() }),
       ]);
