+++ src/testUtils/userMock.ts
@@ -22,7 +22,7 @@ import {
   querySuccessFactory,
 } from "@/testUtils/rtkQueryFactories";
 
-import { userFactory } from "@/testUtils/factories/apiFactories";
+import { userFactory } from "@/testUtils/factories/authFactories";
 
 // Instead of monkey-patching RTK Query's Redux State, we might want to instead encourage using setupRedux
 // in the render helper?
